+++ tools/RAiDER/constants.py
@@ -11,10 +11,14 @@ _ZMIN = np.float64(-100)   # minimum required height
 _ZREF = np.float64(15000)  # maximum requierd height
 _STEP = np.float64(15.0)     # integration step size in meters
 
-_CUBE_SPACING_IN_M = float(2000)  # Horizontal spacing of cube
-
-_g0 = np.float64(9.80665)
+_g0 = np.float64(9.80665) # Standard gravitational constant
+_g1 = np.float64(9.80616) # Gravitational constant @ 45° latitude used for corrections of earth's centrifugal force
 _RE = np.float64(6371008.7714)
 
-R_EARTH_MAX = 6378137
-R_EARTH_MIN = 6356752
+R_EARTH_MAX_WGS84 = 6378137
+R_EARTH_MIN_WGS84 = 6356752
+
+_CUBE_SPACING_IN_M = float(2000)  # Horizontal spacing of cube
+_THRESHOLD_SECONDS = 1 * 60 # Threshold delta_time in seconds
+
+